october tweaks
Fixed a name and bio. Added MetaCTF PDF to files, Added link to it from home page and Day 1 schedule. Day 2 MetaCTF link was added to go to their home page. Added link to Toool.org home page for both days in schedule. Added text on home page to note where the booth map can be obtained, as well as a link to the separate sponsors page - which is nearly the same as the home page data.